requirements.txt
tensorflow>=2.11.0
keras>=2.11.0
matplotlib>=3.5.1
seaborn>=0.11.2
pandas>=1.3.5
numpy>=1.21.5
tqdm>=4.64.0
Pillow>=9.0.0
scikit-learn>=1.0.2
jupyterlab>=3.2.8
streamlit>=1.22.0